STEAM!  :)
I think this will work, although I didn't get to finish testing
everything because I got locked out of Steam for logging in too many
times.  Also, one of the modules is included because I had to manually
rename some of the stuff in it so the two modules could work together.